Start and End Date Fix
Before, I allowed the start and end date to be the same but this caused an issue.

The issue was that you could create multiple terms with the same start and end date if the start and end date are the same.

I remove this so now the start and end dates cannot be the same day.