Update Course and Fixed Bug
Working on the code for the update course screen.

So far, I've created the code to display the list of courses associated with a term.

I also found a bug in my code that checks if the course start and end dates are within range of the terms start and end dates.