Create Course Activity Switch Screen Bug Fix
There was a bug with the cancel button switching screens which would throw an error and crash the application.

Issue: If the user
- goes from the Create Course Activity to the Create Instructor Activity
- cancels on the Create Instructor Activity and goes back to the Create Course Activity
- goes back to the Create Instructor Activity
- cancels on the Create Instructor Activity and goes back to the Create Course Activity
- cancels on the Create Course Activity

The application will end up going to the Create Instructor Activity, which we don't want. Then this could cause an error or an error if the user cancels.

Root Cause: The variable "activityCameFrom" was being overridden with the Create Instructor Activity value.

Instead, I wanted this variable to hold reference to the activity which triggered the Create Course Activity.