Fixed Screen Switch Bug
Issue#1:

Found a bug with the create or update course activity and create or update instructor activity not passing the values needed, (course id and term id), for the screens.

This causes the app to crash.

Issue#2:

Passed in the value for whether the screen that called the create or update instructor activity was for the create or update course.

This was for updating the title of the activity to say either "create course" or "update course"