App Lagging and Crashing Root Cause
The root cause for why my app was crashing and lagging looks to be due to the Thread.sleep(1000) calls in the Repository class.

Specifically, due to all of the database calls and them calling this method.

Note: I might try storing a reference to the database call once to reduce the amount of calls.