Refactored
I removed the multiple database calls for table data some classes and instead have a global variable in the class that retrieves the data once.

This has reduced the amount of unnecessary database calls and slowing down the application; which caused it to crash at times due to the database call wait.

Note: Ideally, would need to look into other alternatives for database and API calls, async code, and how to best write it.